requirements.txt
asgiref==3.11.0
attrs==25.4.0
black==25.12.0
click==8.3.1
Django==6.0
djangorestframework==3.16.1
drf-spectacular==0.29.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.5.1
psycopg2-binary==2.9.11
pytokens==0.3.0
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
uritemplate==4.2.0
